requirements.txt
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
greenlet==1.0.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
Pillow==8.2.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pymongo==3.11.3
pyOpenSSL==20.0.1
python-dotenv==0.17.1
queuelib==1.6.1
requests==2.25.1
Scrapy==2.5.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.11
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urllib3==1.26.4
w3lib==1.22.0
zope.interface==5.4.0
